Move compat module outside of templatetags package
